Continued re-implementation of loading a local model into a new manifest
