+++ modules/ImportLocalModel.py
@@ -1,52 +1,28 @@
-from typing import Set, Callable, Tuple
+from typing import Set
 
 
-from .editing.models import  (  IIIF_TEMP_FORMAT, 
-                                INITIAL_TRANSFORM ,
-                                mimetype_from_extension ,
-                                encode_blender_placement )
-from .editing.transforms import  get_object_placement
+from .editing.models import  mimetype_from_extension , configure_model, walk_object_tree
+from .editing.transforms import Placement
+from .editing.collections import move_object_into_collection
 
 import bpy
 from bpy.props import StringProperty
 from bpy.types import Context, Operator
+from bpy_extras.io_utils import ImportHelper
 
+import pathlib # standard package supporting creating
+               # file-schema url from filesystem path
 import logging
 logger = logging.getLogger("iiif.import_local_model")
 
 
 
             
-class ImportLocalModel(Operator):
+class ImportLocalModel(Operator,  ImportHelper):
     """
     Operator that imports a 3D model into blender
-    
-    precontract:
-    string filepath will be to a local filesystem file
-    string mimetype will be a mimetype of the resource
-        typical value of a mimetyp would be "model/gltf-binary"
-        Its the responsibility of clients that execute this Operator
-        to set the mimetype, based on :
-           -- the value of IIIF format property in an IIIF resource
-           -- the Content-Type from a HTTP header in a network fetch
-           -- 
-        
-    post-contract, on SUCCESS return
-    The single new Blender object that supports having location, rotation, scaling
-    properties will be the active_object. There may be other Blender objects created,
-    particularly if the imported resource gets imported as a multiple number of meshes,
-    in which case this Operator will create a parent structure so that they will all be
-    collectively moved,rotated,scaled.
-    
-    No placement of new objects into Blender collections will be made.
-    
-    The initial Blender location, rotation, scaling will be converted into a json-encoded
-    string as a custom property "iiif_initial_placement" of the active_object
-    
-    Developer Note: This operator does not have invoke method or setup of a UI. In the
-    scenarios at present this operator will be executed from within another Operator 
-    instance that provides an required UI
     """
+
     bl_idname = "iiif.import_local_model"
     bl_label = "Import local file as model"    
     
@@ -65,94 +41,59 @@ class ImportLocalModel(Operator):
         subtype="NONE",
         options={'HIDDEN'}
     )
+
+    # Developer Note 8/19/2025 : Since this class inherits from ImportHelper,
+    # this invoke definition just repeats the built-in inheritance behavior of
+    # calling the invoke method of a superclass.
+    # 
+    # The only reason for adding this definition is to add the informative
+    # logging message.
+    def invoke(self, context, event):
+        logger.info("ImportLocalModel.invoke entered")
+        rv = ImportHelper.invoke(self, context,event)
+        return rv
         
     def execute(self, context: Context) -> Set[str]:
 
-        mimetype = self.mimetype or mimetype_from_extension(self.filepath)
+        self.mimetype = self.mimetype or mimetype_from_extension(self.filepath)
             
         # the key string is a mimetype
         # the value is callable object compatible with
         # the Operator.execute method; it will accept a filepath argument
         # and return set of result values
         
-        # Developer note: this dictionary is constructed at run time
-        # so that an existing importer Operator can be wrapped at run-time
-        # with a wrapper function that supplied other import arguments
-
-        try:
-            handler, handler_name = handler_for_mimetype(mimetype)
-        except KeyError:
-            message = "unsupported mimetype : %s" % mimetype
-            logger.warn(message)
-            return {'CANCELLED'}
-           
-        try:
-            retCode = handler(filepath=self.filepath)
-            if "FINISHED" not in retCode:
-                logger.warn("import handler returned %r"  % (retCode,))
-                return retCode
-            if len(retCode) > 1:
-                logger.info( "import handler returned %r"  % (retCode,) )
-        except Exception as exc:
-            logger.error("glTF import error", exc)
-            return {"CANCELLED"}
-        
-        #new_model = bpy.context.active_object -- potential cruft, remove if when this code works
+        local_loader = bpy.ops.iiif.load_local_model # pyright: ignore[reportAttributeAccessIssue]
+        load_result = local_loader(filepath = self.filepath,  mimetype= self.mimetype)
+        if 'FINISHED' not in load_result:
+            return load_result
+            
         new_model = context.active_object
         if new_model is None:
             logger.warn("context.active_object is None")
             return {"CANCELLED"}
-        else: 
-            # reminder: The IIIF_TEMP_FORMAT value is defined in editing.models
-            # this custom property is defined here and removed by the configure_model
-            # function; it is essentially a way of passing data from this Operator instance
-            # to client code that executes it. 
-            new_model[IIIF_TEMP_FORMAT] = self.mimetype
             
-            
-            blender_transform_encoding : str  = encode_blender_placement(
-                                                    get_object_placement(new_model)
-                                                )
-            logger.debug(f"initial transform: {blender_transform_encoding}")
-            new_model[INITIAL_TRANSFORM] = blender_transform_encoding
-            
+        # configure model by definine IIIF resource properties.
+        # the id will be set based on the local file system path.
+        # if this function is called by the LoadNetworkModel it will
+        # be the responsibility of that client to replace the id
+        # with an id from the network URL
+        model_data = {
+            "id" : pathlib.Path(self.filepath).as_uri(),
+            "format" : self.mimetype,
+            "type"   : "Model"
+        }
+        placement = Placement() # use the identity placement for the new import_scene
+        configure_model(new_model, model_data,placement)
+        
+        LOOP_GUARD_MAX=8
+        for depth, _obj in walk_object_tree(new_model):
+            if depth > LOOP_GUARD_MAX:
+                raise Exception("infinite (or too deep) object parent-child tree")
+            move_object_into_collection(_obj, context.collection)
+                    
         
         
         # properties 
         return {"FINISHED"}
 
 
-def wrapped_gltf(*args, **keyw) -> Set[str]:
-    """
-    A wrapper around the Blender addon-core gltf Importer.
-    this wrapper serves the purpose of quieting the logging INFO messages
-    see:
-    blender/scripts/addons_core/io_scene_gltf2/__init__.py
-    class ImportGLTF2; function set_debug_log
-    """
-    saved_debug_value = bpy.app.debug_value
-    bpy.app.debug_value = 1 # this is equivalent to logging.WARN
-    try:
-        return bpy.ops.import_scene.gltf(*args, **keyw)
-    finally:
-        bpy.app.debug_value = saved_debug_value
-
-def handler_for_mimetype( mimetype:str) -> Tuple[Callable[..., Set[str]] , str] :
-    """
-    return 2-tuple (func, label)
-    function is the callable, may be Blender defined operator call such as 
-    bpy.ops.import_scene.gltf 
-    label a string used only for logging messages
-    
-    """
-    
-    GLTF_IMPORTER : Tuple[Callable[..., Set[str]] , str] = \
-        (wrapped_gltf, "glTF/glb Blender core add-on")
-        
-        
-    mimetype_importer_dict = {
-        "model/gltf-binary" : GLTF_IMPORTER,
-        "model/gltf+json"   : GLTF_IMPORTER,
-    }
-    
-    return mimetype_importer_dict[mimetype]
